Rename the directory from plugin to plugins
