+++ src/flutter/shell/platform/linux_embedded/flutter_linuxes_view.h
@@ -13,12 +13,12 @@
 #include "flutter/shell/platform/embedder/embedder.h"
 #include "flutter/shell/platform/linux_embedded/flutter_linuxes_engine.h"
 #include "flutter/shell/platform/linux_embedded/flutter_linuxes_state.h"
-#include "flutter/shell/platform/linux_embedded/plugin/key_event_plugin.h"
-#include "flutter/shell/platform/linux_embedded/plugin/lifecycle_plugin.h"
-#include "flutter/shell/platform/linux_embedded/plugin/mouse_cursor_plugin.h"
-#include "flutter/shell/platform/linux_embedded/plugin/navigation_plugin.h"
-#include "flutter/shell/platform/linux_embedded/plugin/platform_plugin.h"
-#include "flutter/shell/platform/linux_embedded/plugin/text_input_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/key_event_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/lifecycle_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/mouse_cursor_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/navigation_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/platform_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/text_input_plugin.h"
 #include "flutter/shell/platform/linux_embedded/public/flutter_linuxes.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler_delegate.h"
