+++ src/flutter/shell/platform/linux_embedded/plugins/keyboard_glfw_util.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_KEYBOARD_GLFW_UTIL_H_
-#define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_KEYBOARD_GLFW_UTIL_H_
+#ifndef FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_KEYBOARD_GLFW_UTIL_H_
+#define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_KEYBOARD_GLFW_UTIL_H_
 
 #include <xkbcommon/xkbcommon.h>
 
@@ -17,4 +17,4 @@ uint32_t GetGlfwKeycode(uint32_t xkb_keycode);
 
 }  // namespace flutter
 
-#endif  // FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_KEYBOARD_GLFW_UTIL_H_
+#endif  // FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_KEYBOARD_GLFW_UTIL_H_
