+++ src/flutter/shell/platform/linux_embedded/plugins/navigation_plugin.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_NAVIGATION_PLUGIN_H_
-#define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_NAVIGATION_PLUGIN_H_
+#ifndef FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_NAVIGATION_PLUGIN_H_
+#define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_NAVIGATION_PLUGIN_H_
 
 #include <rapidjson/document.h>
 
@@ -32,4 +32,4 @@ class NavigationPlugin {
 
 }  // namespace flutter
 
-#endif  // FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_NAVIGATION_PLUGIN_H_
+#endif  // FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_NAVIGATION_PLUGIN_H_
