+++ src/flutter/shell/platform/linux_embedded/plugins/platform_plugin.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/linux_embedded/plugin/platform_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/platform_plugin.h"
 
 #include "flutter/shell/platform/common/json_method_codec.h"
