+++ src/flutter/shell/platform/linux_embedded/plugins/text_input_plugin.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "flutter/shell/platform/linux_embedded/plugin/text_input_plugin.h"
+#include "flutter/shell/platform/linux_embedded/plugins/text_input_plugin.h"
 
 #include <linux/input-event-codes.h>
