+++ src/flutter/shell/platform/linux_embedded/plugins/text_input_plugin.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_TEXT_INPUT_PLUGIN_H_
-#define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_TEXT_INPUT_PLUGIN_H_
+#ifndef FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_TEXT_INPUT_PLUGIN_H_
+#define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_TEXT_INPUT_PLUGIN_H_
 
 #include <rapidjson/document.h>
 
@@ -58,4 +58,4 @@ class TextInputPlugin {
 
 }  // namespace flutter
 
-#endif  // FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGIN_TEXT_INPUT_PLUGIN_H_
+#endif  // FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_PLUGINS_TEXT_INPUT_PLUGIN_H_
